Moving the coordination number to mctc-lib (#71)
It would be quite beneficial to have a centralized place to calculate
the CN since it appears in all of our projects (xtb, tblite, dftd4,
dftd3, multicharge, DRACO, ...). Moreover, there are now several
different CNs based on slight variations of the counting function, which
must be integrated in several places.

`mctc-lib` might be the right spot since it is purely based on the
structure, but I am open to alternatives (@awvwgk, @Albkat,
@marvinfriede). As an example, I moved the tblite implementation to the
mctc-lib. This also includes the cutoffs for periodicity and part of the
data sections (maybe we should move the whole as literature radii and EN
are also frequently used).

If we should do this, I will still clean up a few things in the CN (i.e.
rename the gfn CN to dexp for double exponential) to make it general.